[STM32F7] Add syscfg clocking system
