+++ hw/bsp/nucleo-f746zg/syscfg.yml
@@ -59,4 +59,18 @@ syscfg.vals:
     CONFIG_FCB_FLASH_AREA: FLASH_AREA_NFFS
     NFFS_FLASH_AREA: FLASH_AREA_NFFS
     COREDUMP_FLASH_AREA: FLASH_AREA_IMAGE_1
-    MCU_SYSCLK_PLL_HSI: 0
+    STM32_CLOCK_VOLTAGESCALING_CONFIG: 'PWR_REGULATOR_VOLTAGE_SCALE2'
+    STM32_CLOCK_HSI: 0
+    STM32_CLOCK_HSE: 1
+    STM32_CLOCK_HSE_BYPASS: 1
+    STM32_CLOCK_PLL_PLLM: 8
+    STM32_CLOCK_PLL_PLLN: 432
+    STM32_CLOCK_PLL_PLLP: 2
+    STM32_CLOCK_PLL_PLLQ: 9
+    STM32_CLOCK_ENABLE_OVERDRIVE: 0
+    STM32_CLOCK_AHB_DIVIDER: 'RCC_SYSCLK_DIV1'
+    STM32_CLOCK_APB1_DIVIDER: 'RCC_HCLK_DIV4'
+    STM32_CLOCK_APB2_DIVIDER: 'RCC_HCLK_DIV2'
+    STM32_FLASH_LATENCY: 'FLASH_LATENCY_7'
+    STM32_FLASH_PREFETCH_ENABLE: 1
+    STM32_ART_ACCLERATOR_ENABLE: 1
