+++ hw/bsp/stm32f7discovery/src/hal_bsp.c
@@ -178,8 +178,6 @@ hal_bsp_init(void)
 
     (void)rc;
 
-    hal_system_clock_start();
-
 #if MYNEWT_VAL(UART_0)
     rc = os_dev_create((struct os_dev *) &hal_uart0, "uart0",
       OS_DEV_INIT_PRIMARY, 0, uart_hal_init, (void *)&uart_cfg[0]);
